refactor dir and add latte
