+++ train.py
@@ -24,8 +24,8 @@ from torch.utils.tensorboard import SummaryWriter
 from torchvision.datasets import CIFAR10
 from tqdm import tqdm
 
-from opendit.models.diffusion import create_diffusion
-from opendit.models.dit import DiT, DiT_models
+from opendit.diffusion import create_diffusion
+from opendit.dit.dit import DiT, DiT_models
 from opendit.utils.ckpt_utils import create_logger, load, record_model_param_shape, save
 from opendit.utils.data_utils import prepare_dataloader
 from opendit.utils.operation import model_sharding
